Cleaned bbswitch and switcheroo switchers code
The code for writing ON/OFF was equal except for the string to be written. To
avoid code duplication, introduce a new static *_write function. Also, the path
to bbswitch and switcheroo are now macros to make future changes to that easier.
The debugging messages for opening/writing has been extended to include more
debugging information.