Detect loaded driver from PCI Bus ID
This method is more reliable than trying to unload a set of drivers or checking
for them. For example, the old driver "nv" (is anybody still using them? ugh)
would not get caught by the previous check and the card would be attempted to be
disabled anyway. The new check matches the code of the Bumblebee 2.4 (in Bash)
by looking up drivers by matching the PCI Bus ID.