Improve detection code for bbswitch and vga_switcheroo (GH-13)
The validation for bbswitch includes a basic check: if it's loaded it is
available. However, if it's not loaded, it is not sure whether the module is
usable or not, this depends on whether the card is supported or not.
vga_switcheroo is more tricky. Even if the file exists, it does not necessarily
mean that the card is supported.