Drop ENABLE_POWER_MANAGEMENT, support PM_METHOD (Closes GH-14)
ENABLE_POWER_MANAGEMENT was used solely for the apci_call method. Dropping this
as we now have other methods. In addition, debugging messages now show the name
of the chosen settings instead of a number. The PM_METHOD setting for using
vga_switcheroo has been changed from "vgaswitcheroo" to "switcheroo"