Base driver detection on the available kernel modules
If the compile default for DRIVER is set and no override has been found on the
command line nor config file, then it will be used. Otherwise, autodetection is
performed on the loaded drivers (currently nvidia and nouveau, in that order
because if the nvidia blob is installed, the user likely wants to use that.)
If that failed, the driver will be looked up through modinfo.