Fill in configure defaults into bumblebee.conf
Added a function to bbconfig.c for converting a PM method string to an internal
representation of it (an int from the enum). This is needed for being able to
fill in "auto" in the conffile. To be able to expand $sysconfdir correctly in
bumblebee.conf, shell expansion had to occur. This has as a side-effect that
quotes are removed, which could also be seen as a positive thing.