Fix -Wunused-parameter warnings (Closes GH-210)
In the past, configuration functions needed argc/argv, but this is not
true these days. Get rid of it.

In sw_bbswitch.c, one could check whether the card is a nvidia one, but
bbswitch does a good job at reporting errors when detection fails, so
just ignore the info parameter.