Adds a fix for two nvidia cards (similar to Optimus technology).
Now it seems to be working on a Macbook Pro 5,1 with the following VGA:
  * GeForce 9400 M (int.)
  * GeForce 9600 M GT (dis.)

Changes pci_find_gfx_by_vendor's interface so it now uses a numeric
index, to keep track of both graphic cards if they are from the same
vendor.

Besides, we now print (with LOG_DEBUG level) the PCI Bus ID of each card
and how we categorized them (discrete/integrated).