module: support module aliases
As reported on GH-330, bumblebeed bails out when the KernelDriver line is set to
"nvidia", even though an alias exist in /etc/modprobe.d/. Instead of using
`modinfo` to determine the existence of a module, use `modprobe --dry-run` to
determine whether a module is available.

This should also catch errors earlier when `nouveau` is aliased to `off` for
example in which case the module must be considered "unavailable".