Fix module unloading
On Ubuntu, "modprobe nvidia" does not work if there is no such alias.
Previously this code ran "rmmod nvidia", but that broke since the
introduction of nvidia-uvm.

This patch attempts to map "nvidia" to "nvidia-304" by querying the
configuration and is not foolproof (e.g. when nouveau is configured
instead). As a workaround, you can create a modprobe.conf file
containing:

    alias nvidia nvidia-304